# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Task API",
    "version": "1.0.0",
    "description": "A simple CRUD API for managing tasks"
  },
  "paths": {
  "/tasks": {
    "get": {
      "summary": "Get all tasks",
      "responses": {
        "200": {
          "description": "List of all tasks"
        }
      }
    },

    "post": {
      "summary": "Create a new task",
      "requestBody": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": ["title"],
              "properties": {
                "title": {
                  "type": "string",
                  "example": "Buy milk"
                }
              }
            }
          }
        }
      },
      "responses": {
        "201": {
          "description": "Task created"
        },
        "400": {
          "description": "Invalid request"
        }
      }
    }
  },

  "/tasks/{id}": {
    "get": {
      "summary": "Get a task by ID",
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer"
          }
        }
      ],
      "responses": {
        "200": {
          "description": "Task found"
        },
        "404": {
          "description": "Task not found"
        }
      }
    },

    "put": {
      "summary": "Update a task",
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer"
          }
        }
      ],
      "requestBody": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string",
                  "example": "Learn Express"
                },
                "done": {
                  "type": "boolean",
                  "example": true
                }
              }
            }
          }
        }
      },
      "responses": {
        "200": {
          "description": "Task updated"
        },
        "400": {
          "description": "Invalid request"
        },
        "404": {
          "description": "Task not found"
        }
      }
    },

    "delete": {
      "summary": "Delete a task",
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer"
          }
        }
      ],
      "responses": {
        "204": {
          "description": "Task deleted"
        },
        "404": {
          "description": "Task not found"
        }
      }
    }
  }
}
}